test: update test setup
