--- src/__tests__/mockAvatar.png
@@ -1,1 +1,0 @@
-fake file content
